+++ NEWS.md
@@ -1,3 +1,8 @@
-# binaryLogic 0.3 
+# binaryLogic 0.3.3
 
-* Added new attribute »stickyBits« to switchEndianess. See `?switchEndianess()` for more details.
+* Added a new attribute size n [Bit] to »as.binary« for unsigned number. See `?as.binary()`
+* Converter to gray code. See `?bin2gray()` and `?gray2bin()`
+
+# binaryLogic 0.3.2
+
+* Added a new attribute »stickyBits« to switchEndianess. See `?switchEndianess()` for more details.